Fixed the file structure, Added the SMTPCLASS and made update in the readme
