+++ .gitignore
@@ -131,6 +131,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
+/env/
 .env
 .python-version
 .venv
