+++ compose.yaml
@@ -40,3 +40,5 @@ services:
         - path: src/streamlit_app.py
           action: sync+restart
           target: /app/streamlit_app.py
+
+
