+++ requirements.txt
@@ -1,7 +1,128 @@
-# IMPORTANT: This is only used for the Streamlit app deployment
-# to Streamlit Community Cloud, since it doesn't have support for uv.
-# USE PYPROJECT.TOML AND UV.LOCK FOR YOUR OWN INSTALL!
-
-httpx~=0.26.0
-pydantic ~=2.9.0
-streamlit~=1.37.0
+aiohappyeyeballs==2.4.3
+aiohttp==3.11.7
+aiosignal==1.3.1
+aiosmtplib==2.0.2
+aiosqlite==0.20.0
+altair==5.4.1
+annotated-types==0.7.0
+anthropic==0.39.0
+anyio==4.6.2.post1
+attrs==24.2.0
+blinker==1.9.0
+boto3==1.35.66
+botocore==1.35.66
+cachetools==5.5.0
+certifi==2024.8.30
+charset-normalizer==3.4.0
+click==8.1.7
+colorama==0.4.6
+dataclasses-json==0.6.7
+defusedxml==0.7.1
+distro==1.9.0
+dnspython==2.7.0
+duckduckgo_search==6.3.5
+email_validator==2.2.0
+fastapi==0.115.5
+fastapi-mail==1.4.1
+frozenlist==1.5.0
+gitdb==4.0.11
+GitPython==3.1.43
+google-ai-generativelanguage==0.6.10
+google-api-core==2.23.0
+google-api-python-client==2.154.0
+google-auth==2.36.0
+google-auth-httplib2==0.2.0
+google-generativeai==0.8.3
+googleapis-common-protos==1.66.0
+greenlet==3.1.1
+groq==0.12.0
+grpcio==1.68.0
+grpcio-status==1.68.0
+h11==0.14.0
+httpcore==1.0.7
+httplib2==0.22.0
+httpx==0.27.2
+httpx-sse==0.4.0
+idna==3.10
+itsdangerous==2.2.0
+Jinja2==3.1.4
+jiter==0.7.1
+jmespath==1.0.1
+jsonpatch==1.33
+jsonpointer==3.0.0
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+langchain==0.3.7
+langchain-anthropic==0.3.0
+langchain-aws==0.2.7
+langchain-community==0.3.7
+langchain-core==0.3.19
+langchain-google-genai==2.0.5
+langchain-groq==0.2.1
+langchain-openai==0.2.9
+langchain-text-splitters==0.3.2
+langgraph==0.2.52
+langgraph-checkpoint==2.0.5
+langgraph-checkpoint-sqlite==2.0.1
+langgraph-sdk==0.1.36
+langsmith==0.1.143
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+marshmallow==3.23.1
+mdurl==0.1.2
+msgpack==1.1.0
+multidict==6.1.0
+mypy-extensions==1.0.0
+narwhals==1.14.1
+numexpr==2.10.1
+numpy==1.26.4
+openai==1.55.0
+orjson==3.10.11
+packaging==24.2
+pandas==2.2.3
+pillow==11.0.0
+primp==0.7.0
+propcache==0.2.0
+proto-plus==1.25.0
+protobuf==5.28.3
+pyarrow==18.0.0
+pyasn1==0.6.1
+pyasn1_modules==0.4.1
+pydantic==2.9.2
+pydantic-settings==2.6.1
+pydantic_core==2.23.4
+pydeck==0.9.1
+Pygments==2.18.0
+pyparsing==3.2.0
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+python-multipart==0.0.17
+pytz==2024.2
+PyYAML==6.0.2
+referencing==0.35.1
+regex==2024.11.6
+requests==2.32.3
+requests-toolbelt==1.0.0
+rich==13.9.4
+rpds-py==0.21.0
+rsa==4.9
+s3transfer==0.10.4
+six==1.16.0
+smmap==5.0.1
+sniffio==1.3.1
+SQLAlchemy==2.0.35
+starlette==0.41.3
+streamlit==1.40.1
+tenacity==9.0.0
+tiktoken==0.8.0
+toml==0.10.2
+tornado==6.4.1
+tqdm==4.67.0
+typing-inspect==0.9.0
+typing_extensions==4.12.2
+tzdata==2024.2
+uritemplate==4.1.1
+urllib3==2.2.3
+uvicorn==0.32.0
+watchdog==6.0.0
+yarl==1.18.0
