+++ src/main/__init__.py
@@ -1,0 +1,3 @@
+from main.main import app
+
+__all__ = ["app"]
