--- src/service/__init__.py
@@ -1,3 +1,0 @@
-from service.service import app
-
-__all__ = ["app"]
